requirements.txt
absl-py==0.11.0
argon2-cffi==20.1.0
asteval==0.9.21
astor==0.8.1
async-generator==1.10
atari-py==0.2.6
attrs==20.3.0
autopep8==1.5.4
backcall==0.2.0
bleach==3.2.1
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cloudpickle==1.6.0
colorama==0.4.4
cycler==0.10.0
darkgreybox==0.1.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
gast==0.4.0
google-auth==1.24.0
google-pasta==0.2.0
grpcio==1.34.0
gurobipy==9.1.1
gym==0.18.0
h5py==3.1.0
idna==2.10
importlib-metadata==3.3.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.2
jedi==0.18.0
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lmfit==1.0.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
mpi4py==3.0.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
nose==1.3.7
notebook==6.1.5
numdifftools==0.9.39
numpy==1.19.3
oauthlib==3.1.0
opencv-python==4.5.1.48
packaging==20.8
pandas==1.1.2
pandocfilters==1.4.3
parso==0.8.1
patsy==0.5.1
pickleshare==0.7.5
Pillow==7.2.0
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyglet==1.5.0
Pygments==2.7.3
Pyomo==5.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
PyUtilib==6.0.0
pywin32==300
pywinpty==0.5.7
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.6.0
Send2Trash==1.5.0
six==1.15.0
stable-baselines==2.10.1
statsmodels==0.11.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
uncertainties==3.1.5
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.0
